Move lots of free functions to `TestConfig` methods
