+++ src/lib.rs
@@ -13,18 +13,16 @@ pub use color_eyre;
 use color_eyre::eyre::{eyre, Result};
 use crossbeam_channel::{unbounded, Receiver, Sender};
 use dependencies::{Build, BuildManager};
-use parser::{ErrorMatch, ErrorMatchKind, OptWithLine, Revisioned, Spanned};
 use per_test_config::TestConfig;
-use rustc_stderr::{Level, Message};
-use spanned::Span;
+use rustc_stderr::Message;
 use status_emitter::{StatusEmitter, TestStatus};
-use std::collections::{HashSet, VecDeque};
+use std::collections::VecDeque;
 use std::ffi::OsString;
 use std::num::NonZeroUsize;
-use std::path::{Component, Path, PathBuf, Prefix};
+use std::path::{Component, Path, Prefix};
 use std::process::{Command, Output};
 use std::thread;
-use test_result::{Errored, TestOk, TestResult, TestRun};
+use test_result::{Errored, TestOk, TestRun};
 
 use crate::parser::Comments;
 
@@ -425,7 +423,7 @@ fn parse_and_test_file(
                 path: status.path(),
             };
 
-            let result = run_test(build_manager, test_config);
+            let result = test_config.run_test(build_manager);
             TestRun { result, status }
         })
         .collect())
@@ -490,7 +488,7 @@ fn build_aux(
 
     let mut aux_cmd = config.build_command()?;
 
-    let mut extra_args = build_aux_files(aux_file.parent().unwrap(), &config, build_manager)?;
+    let mut extra_args = config.build_aux_files(aux_file.parent().unwrap(), build_manager)?;
     // Make sure we see our dependencies
     aux_cmd.args(extra_args.iter());
 
@@ -531,43 +529,6 @@ fn build_aux(
     Ok(extra_args)
 }
 
-fn run_test(build_manager: &BuildManager<'_>, mut config: TestConfig<'_>) -> TestResult {
-    let extra_args = build_aux_files(
-        &config.path.parent().unwrap().join("auxiliary"),
-        &config,
-        build_manager,
-    )?;
-
-    config.patch_out_dir();
-
-    let mut cmd = config.build_command()?;
-    cmd.args(&extra_args);
-    let stdin = config.path.with_extension(config.extension("stdin"));
-    if stdin.exists() {
-        cmd.stdin(std::fs::File::open(stdin).unwrap());
-    }
-
-    let (cmd, output) = run_command(cmd)?;
-
-    let mode = config.mode()?;
-    let (cmd, output) = check_test_result(
-        cmd,
-        match *mode {
-            Mode::Run { .. } => Mode::Pass,
-            _ => *mode,
-        },
-        &config,
-        output,
-    )?;
-
-    if let Mode::Run { .. } = *mode {
-        run_test_binary(mode, cmd, config)
-    } else {
-        run_rustfix(output, config, *mode, extra_args)?;
-        Ok(TestOk::Ok)
-    }
-}
-
 fn run_command(mut cmd: Command) -> Result<(Command, Output), Errored> {
     match cmd.output() {
         Err(err) => Err(Errored {
@@ -580,512 +541,6 @@ fn run_command(mut cmd: Command) -> Result<(Command, Output), Errored> {
     }
 }
 
-fn build_aux_files(
-    aux_dir: &Path,
-    config: &TestConfig,
-    build_manager: &BuildManager<'_>,
-) -> Result<Vec<OsString>, Errored> {
-    let mut extra_args = vec![];
-    for rev in config.comments() {
-        for aux in &rev.aux_builds {
-            let line = aux.line();
-            let aux = &**aux;
-            let aux_file = if aux.starts_with("..") {
-                aux_dir.parent().unwrap().join(aux)
-            } else {
-                aux_dir.join(aux)
-            };
-            extra_args.extend(
-                build_manager
-                    .build(
-                        Build::Aux {
-                            aux_file: strip_path_prefix(
-                                &aux_file.canonicalize().map_err(|err| Errored {
-                                    command: Command::new(format!(
-                                        "canonicalizing path `{}`",
-                                        aux_file.display()
-                                    )),
-                                    errors: vec![],
-                                    stderr: err.to_string().into_bytes(),
-                                    stdout: vec![],
-                                })?,
-                                &std::env::current_dir().unwrap(),
-                            )
-                            .collect(),
-                        },
-                        &config.config,
-                    )
-                    .map_err(
-                        |Errored {
-                             command,
-                             errors,
-                             stderr,
-                             stdout,
-                         }| Errored {
-                            command,
-                            errors: vec![Error::Aux {
-                                path: aux_file,
-                                errors,
-                                line,
-                            }],
-                            stderr,
-                            stdout,
-                        },
-                    )?,
-            );
-        }
-    }
-    Ok(extra_args)
-}
-
-fn run_test_binary(mode: Spanned<Mode>, mut cmd: Command, config: TestConfig) -> TestResult {
-    let revision = config.extension("run");
-    let config = TestConfig {
-        config: config.config,
-        revision: &revision,
-        comments: config.comments,
-        path: config.path,
-    };
-    cmd.arg("--print").arg("file-names");
-    let output = cmd.output().unwrap();
-    assert!(output.status.success());
-
-    let mut files = output.stdout.lines();
-    let file = files.next().unwrap();
-    assert_eq!(files.next(), None);
-    let file = std::str::from_utf8(file).unwrap();
-    let exe_file = config.config.out_dir.join(file);
-    let mut exe = Command::new(&exe_file);
-    let stdin = config.path.with_extension(format!("{revision}.stdin"));
-    if stdin.exists() {
-        exe.stdin(std::fs::File::open(stdin).unwrap());
-    }
-    let output = exe
-        .output()
-        .unwrap_or_else(|err| panic!("exe file: {}: {err}", exe_file.display()));
-
-    let mut errors = vec![];
-
-    check_test_output(&mut errors, &config, &output.stdout, &output.stderr);
-
-    errors.extend(mode.ok(output.status).err());
-    if errors.is_empty() {
-        Ok(TestOk::Ok)
-    } else {
-        Err(Errored {
-            command: exe,
-            errors,
-            stderr: vec![],
-            stdout: vec![],
-        })
-    }
-}
-
-fn run_rustfix(
-    output: Output,
-    config: TestConfig,
-    mode: Mode,
-    extra_args: Vec<OsString>,
-) -> Result<(), Errored> {
-    let no_run_rustfix = config.find_one("`no-rustfix` annotations", |r| r.no_rustfix.clone())?;
-
-    let global_rustfix = match mode {
-        Mode::Pass | Mode::Run { .. } | Mode::Panic => RustfixMode::Disabled,
-        Mode::Fail { rustfix, .. } | Mode::Yolo { rustfix } => rustfix,
-    };
-
-    let fixed_code = (no_run_rustfix.is_none() && global_rustfix.enabled())
-        .then_some(())
-        .and_then(|()| {
-            let suggestions = std::str::from_utf8(&output.stderr)
-                .unwrap()
-                .lines()
-                .flat_map(|line| {
-                    if !line.starts_with('{') {
-                        return vec![];
-                    }
-                    rustfix::get_suggestions_from_json(
-                        line,
-                        &HashSet::new(),
-                        if global_rustfix == RustfixMode::Everything {
-                            rustfix::Filter::Everything
-                        } else {
-                            rustfix::Filter::MachineApplicableOnly
-                        },
-                    )
-                    .unwrap_or_else(|err| {
-                        panic!("could not deserialize diagnostics json for rustfix {err}:{line}")
-                    })
-                })
-                .collect::<Vec<_>>();
-            if suggestions.is_empty() {
-                None
-            } else {
-                let path_str = config.path.display().to_string();
-                for sugg in &suggestions {
-                    for snip in &sugg.snippets {
-                        if snip.file_name != path_str {
-                            return Some(Err(anyhow::anyhow!("cannot apply suggestions for `{}` since main file is `{path_str}`. Please use `//@no-rustfix` to disable rustfix", snip.file_name)));
-                        }
-                    }
-                }
-                Some(rustfix::apply_suggestions(
-                    &std::fs::read_to_string(config.path).unwrap(),
-                    &suggestions,
-                ))
-            }
-        })
-        .transpose()
-        .map_err(|err| Errored {
-            command: Command::new(format!("rustfix {}", config.path.display())),
-            errors: vec![Error::Rustfix(err)],
-            stderr: output.stderr,
-            stdout: output.stdout,
-        })?;
-
-    let edition = config.edition()?.into();
-    let rustfix_comments = Comments {
-        revisions: None,
-        revisioned: std::iter::once((
-            vec![],
-            Revisioned {
-                span: Span::default(),
-                ignore: vec![],
-                only: vec![],
-                stderr_per_bitwidth: false,
-                compile_flags: config.collect(|r| r.compile_flags.iter().cloned()),
-                env_vars: config.collect(|r| r.env_vars.iter().cloned()),
-                normalize_stderr: vec![],
-                normalize_stdout: vec![],
-                error_in_other_files: vec![],
-                error_matches: vec![],
-                require_annotations_for_level: Default::default(),
-                aux_builds: config.collect(|r| r.aux_builds.iter().cloned()),
-                edition,
-                mode: OptWithLine::new(Mode::Pass, Span::default()),
-                no_rustfix: OptWithLine::new((), Span::default()),
-                diagnostic_code_prefix: OptWithLine::new(String::new(), Span::default()),
-                needs_asm_support: false,
-            },
-        ))
-        .collect(),
-    };
-    let config = TestConfig {
-        config: config.config,
-        revision: config.revision,
-        comments: &rustfix_comments,
-        path: config.path,
-    };
-
-    let run = fixed_code.is_some();
-    let mut errors = vec![];
-    let rustfix_path = check_output(
-        // Always check for `.fixed` files, even if there were reasons not to run rustfix.
-        // We don't want to leave around stray `.fixed` files
-        fixed_code.unwrap_or_default().as_bytes(),
-        &mut errors,
-        "fixed",
-        &config,
-    );
-    // picking the crate name from the file name is problematic when `.revision_name` is inserted,
-    // so we compute it here before replacing the path.
-    let crate_name = config
-        .path
-        .file_stem()
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .replace('-', "_");
-    let config = TestConfig {
-        config: config.config,
-        revision: config.revision,
-        comments: &rustfix_comments,
-        path: &rustfix_path,
-    };
-    if !errors.is_empty() {
-        return Err(Errored {
-            command: Command::new(format!("checking {}", config.path.display())),
-            errors,
-            stderr: vec![],
-            stdout: vec![],
-        });
-    }
-
-    if !run {
-        return Ok(());
-    }
-
-    let mut cmd = config.build_command()?;
-    cmd.args(extra_args);
-    cmd.arg("--crate-name").arg(crate_name);
-    let output = cmd.output().unwrap();
-    if output.status.success() {
-        Ok(())
-    } else {
-        Err(Errored {
-            command: cmd,
-            errors: vec![Error::Command {
-                kind: "rustfix".into(),
-                status: output.status,
-            }],
-            stderr: rustc_stderr::process(&rustfix_path, &output.stderr).rendered,
-            stdout: output.stdout,
-        })
-    }
-}
-
-fn check_test_result(
-    command: Command,
-    mode: Mode,
-    config: &TestConfig,
-    output: Output,
-) -> Result<(Command, Output), Errored> {
-    let mut errors = vec![];
-    errors.extend(mode.ok(output.status).err());
-    // Always remove annotation comments from stderr.
-    let diagnostics = rustc_stderr::process(config.path, &output.stderr);
-    check_test_output(&mut errors, config, &output.stdout, &diagnostics.rendered);
-    // Check error annotations in the source against output
-    check_annotations(
-        diagnostics.messages,
-        diagnostics.messages_from_unknown_file_or_line,
-        config,
-        &mut errors,
-    )?;
-    if errors.is_empty() {
-        Ok((command, output))
-    } else {
-        Err(Errored {
-            command,
-            errors,
-            stderr: diagnostics.rendered,
-            stdout: output.stdout,
-        })
-    }
-}
-
-fn check_test_output(errors: &mut Vec<Error>, config: &TestConfig, stdout: &[u8], stderr: &[u8]) {
-    // Check output files (if any)
-    // Check output files against actual output
-    check_output(stderr, errors, "stderr", config);
-    check_output(stdout, errors, "stdout", config);
-}
-
-fn check_annotations(
-    mut messages: Vec<Vec<Message>>,
-    mut messages_from_unknown_file_or_line: Vec<Message>,
-    config: &TestConfig,
-    errors: &mut Errors,
-) -> Result<(), Errored> {
-    let error_patterns = config
-        .comments()
-        .flat_map(|r| r.error_in_other_files.iter());
-
-    let mut seen_error_match = None;
-    for error_pattern in error_patterns {
-        seen_error_match = Some(error_pattern.span());
-        // first check the diagnostics messages outside of our file. We check this first, so that
-        // you can mix in-file annotations with //@error-in-other-file annotations, even if there is overlap
-        // in the messages.
-        if let Some(i) = messages_from_unknown_file_or_line
-            .iter()
-            .position(|msg| error_pattern.matches(&msg.message))
-        {
-            messages_from_unknown_file_or_line.remove(i);
-        } else {
-            errors.push(Error::PatternNotFound {
-                pattern: error_pattern.clone(),
-                expected_line: None,
-            });
-        }
-    }
-    let diagnostic_code_prefix = config
-        .find_one("diagnostic_code_prefix", |r| {
-            r.diagnostic_code_prefix.clone()
-        })?
-        .into_inner()
-        .map(|s| s.content)
-        .unwrap_or_default();
-
-    // The order on `Level` is such that `Error` is the highest level.
-    // We will ensure that *all* diagnostics of level at least `lowest_annotation_level`
-    // are matched.
-    let mut lowest_annotation_level = Level::Error;
-    'err: for &ErrorMatch { ref kind, line } in
-        config.comments().flat_map(|r| r.error_matches.iter())
-    {
-        match kind {
-            ErrorMatchKind::Code(code) => {
-                seen_error_match = Some(code.span());
-            }
-            &ErrorMatchKind::Pattern { ref pattern, level } => {
-                seen_error_match = Some(pattern.span());
-                // If we found a diagnostic with a level annotation, make sure that all
-                // diagnostics of that level have annotations, even if we don't end up finding a matching diagnostic
-                // for this pattern.
-                if lowest_annotation_level > level {
-                    lowest_annotation_level = level;
-                }
-            }
-        }
-
-        if let Some(msgs) = messages.get_mut(line.get()) {
-            match kind {
-                &ErrorMatchKind::Pattern { ref pattern, level } => {
-                    let found = msgs
-                        .iter()
-                        .position(|msg| pattern.matches(&msg.message) && msg.level == level);
-                    if let Some(found) = found {
-                        msgs.remove(found);
-                        continue;
-                    }
-                }
-                ErrorMatchKind::Code(code) => {
-                    for (i, msg) in msgs.iter().enumerate() {
-                        if msg.level != Level::Error {
-                            continue;
-                        }
-                        let Some(msg_code) = &msg.code else { continue };
-                        let Some(msg) = msg_code.strip_prefix(&diagnostic_code_prefix) else {
-                            continue;
-                        };
-                        if msg == **code {
-                            msgs.remove(i);
-                            continue 'err;
-                        }
-                    }
-                }
-            }
-        }
-
-        errors.push(match kind {
-            ErrorMatchKind::Pattern { pattern, .. } => Error::PatternNotFound {
-                pattern: pattern.clone(),
-                expected_line: Some(line),
-            },
-            ErrorMatchKind::Code(code) => Error::CodeNotFound {
-                code: Spanned::new(format!("{}{}", diagnostic_code_prefix, **code), code.span()),
-                expected_line: Some(line),
-            },
-        });
-    }
-
-    let required_annotation_level = config
-        .find_one("`require_annotations_for_level` annotations", |r| {
-            r.require_annotations_for_level.clone()
-        })?;
-
-    let required_annotation_level = required_annotation_level
-        .into_inner()
-        .map_or(lowest_annotation_level, |l| *l);
-    let filter = |mut msgs: Vec<Message>| -> Vec<_> {
-        msgs.retain(|msg| msg.level >= required_annotation_level);
-        msgs
-    };
-
-    let mode = config.mode()?;
-
-    if !matches!(*mode, Mode::Yolo { .. }) {
-        let messages_from_unknown_file_or_line = filter(messages_from_unknown_file_or_line);
-        if !messages_from_unknown_file_or_line.is_empty() {
-            errors.push(Error::ErrorsWithoutPattern {
-                path: None,
-                msgs: messages_from_unknown_file_or_line,
-            });
-        }
-
-        for (line, msgs) in messages.into_iter().enumerate() {
-            let msgs = filter(msgs);
-            if !msgs.is_empty() {
-                let line = NonZeroUsize::new(line).expect("line 0 is always empty");
-                errors.push(Error::ErrorsWithoutPattern {
-                    path: Some(Spanned::new(
-                        config.path.to_path_buf(),
-                        spanned::Span {
-                            line_start: line,
-                            ..spanned::Span::default()
-                        },
-                    )),
-                    msgs,
-                });
-            }
-        }
-    }
-
-    match (*mode, seen_error_match) {
-        (Mode::Pass, Some(span)) | (Mode::Panic, Some(span)) => {
-            errors.push(Error::PatternFoundInPassTest {
-                mode: mode.span(),
-                span,
-            })
-        }
-        (
-            Mode::Fail {
-                require_patterns: true,
-                ..
-            },
-            None,
-        ) => errors.push(Error::NoPatternsFound),
-        _ => {}
-    }
-    Ok(())
-}
-
-fn check_output(
-    output: &[u8],
-    errors: &mut Errors,
-    kind: &'static str,
-    config: &TestConfig,
-) -> PathBuf {
-    let output = normalize(output, config, kind);
-    let path = output_path(config, kind);
-    match &config.config.output_conflict_handling {
-        OutputConflictHandling::Error => {
-            let expected_output = std::fs::read(&path).unwrap_or_default();
-            if output != expected_output {
-                errors.push(Error::OutputDiffers {
-                    path: path.clone(),
-                    actual: output.clone(),
-                    expected: expected_output,
-                    bless_command: config.config.bless_command.clone(),
-                });
-            }
-        }
-        OutputConflictHandling::Bless => {
-            if output.is_empty() {
-                let _ = std::fs::remove_file(&path);
-            } else {
-                std::fs::write(&path, &output).unwrap();
-            }
-        }
-        OutputConflictHandling::Ignore => {}
-    }
-    path
-}
-
-fn output_path(config: &TestConfig<'_>, kind: &str) -> PathBuf {
-    let ext = config.extension(kind);
-    if config.comments().any(|r| r.stderr_per_bitwidth) {
-        return config
-            .path
-            .with_extension(format!("{}bit.{ext}", config.config.get_pointer_width()));
-    }
-    config.path.with_extension(ext)
-}
-
-fn normalize(text: &[u8], config: &TestConfig, kind: &'static str) -> Vec<u8> {
-    let mut text = text.to_owned();
-
-    for (from, to) in config.comments().flat_map(|r| match kind {
-        "fixed" => &[] as &[_],
-        "stderr" => &r.normalize_stderr,
-        "stdout" => &r.normalize_stdout,
-        _ => unreachable!(),
-    }) {
-        text = from.replace_all(&text, to).into_owned();
-    }
-    text
-}
 /// Remove the common prefix of this path and the `root_dir`.
 fn strip_path_prefix<'a>(path: &'a Path, prefix: &Path) -> impl Iterator<Item = Component<'a>> {
     let mut components = path.components();
