+++ src/tests.rs
@@ -1,5 +1,7 @@
 use std::path::{Path, PathBuf};
 
+use spanned::Spanned;
+
 use crate::rustc_stderr::Level;
 use crate::rustc_stderr::Message;
 
@@ -50,7 +52,9 @@ fn main() {
             }
         ]
     ];
-    check_annotations(messages, vec![], &config, &mut errors).unwrap();
+    config
+        .check_annotations(messages, vec![], &mut errors)
+        .unwrap();
     match &errors[..] {
         [Error::PatternNotFound { pattern, .. }, Error::ErrorsWithoutPattern { path, .. }]
             if path.as_ref().is_some_and(|p| p.line().get() == 5) && pattern.line().get() == 5 => {}
@@ -80,7 +84,9 @@ fn main() {
             ]
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [] => {}
             _ => panic!("{:#?}", errors),
@@ -99,7 +105,9 @@ fn main() {
             ]
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [Error::PatternNotFound { pattern, .. }, Error::ErrorsWithoutPattern { path, .. }]
                 if path.as_ref().is_some_and(|p| p.line().get() == 4)
@@ -122,7 +130,9 @@ fn main() {
             ]
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             // Note no `ErrorsWithoutPattern`, because there are no `//~NOTE` in the test file, so we ignore them
             [Error::PatternNotFound { pattern, .. }] if pattern.line().get() == 5 => {}
@@ -155,7 +165,9 @@ fn main() {
         ]
     ];
     let mut errors = vec![];
-    check_annotations(messages, vec![], &config, &mut errors).unwrap();
+    config
+        .check_annotations(messages, vec![], &mut errors)
+        .unwrap();
     match &errors[..] {
         [Error::PatternNotFound { pattern, .. }] if pattern.line().get() == 6 => {}
         _ => panic!("{:#?}", errors),
@@ -191,7 +203,9 @@ fn main() {
         ]
     ];
     let mut errors = vec![];
-    check_annotations(messages, vec![], &config, &mut errors).unwrap();
+    config
+        .check_annotations(messages, vec![], &mut errors)
+        .unwrap();
     match &errors[..] {
         [Error::ErrorsWithoutPattern { path, .. }]
             if path.as_ref().is_some_and(|p| p.line().get() == 5) => {}
@@ -239,7 +253,9 @@ fn main() {
         ],
     ];
     let mut errors = vec![];
-    check_annotations(messages, vec![], &config, &mut errors).unwrap();
+    config
+        .check_annotations(messages, vec![], &mut errors)
+        .unwrap();
     match &errors[..] {
         [Error::ErrorsWithoutPattern { path, msgs, .. }]
             if path.as_ref().is_some_and(|p| p.line().get() == 5) =>
@@ -298,7 +314,9 @@ fn main() {
         ],
     ];
     let mut errors = vec![];
-    check_annotations(messages, vec![], &config, &mut errors).unwrap();
+    config
+        .check_annotations(messages, vec![], &mut errors)
+        .unwrap();
     match &errors[..] {
         [] => {}
         _ => panic!("{:#?}", errors),
@@ -327,7 +345,9 @@ fn main() {
             }],
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [] => {}
             _ => panic!("{:#?}", errors),
@@ -348,7 +368,9 @@ fn main() {
             }],
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [Error::CodeNotFound { code, .. }, Error::ErrorsWithoutPattern { msgs, .. }]
                 if **code == "E0308" && code.line().get() == 3 && msgs.len() == 1 => {}
@@ -370,7 +392,9 @@ fn main() {
             }],
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [Error::CodeNotFound { code, .. }] if **code == "E0308" && code.line().get() == 3 => {}
             _ => panic!("{:#?}", errors),
@@ -402,7 +426,9 @@ fn main() {
             }],
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [] => {}
             _ => panic!("{:#?}", errors),
@@ -423,7 +449,9 @@ fn main() {
             }],
         ];
         let mut errors = vec![];
-        check_annotations(messages, vec![], &config, &mut errors).unwrap();
+        config
+            .check_annotations(messages, vec![], &mut errors)
+            .unwrap();
         match &errors[..] {
             [Error::CodeNotFound { code, .. }, Error::ErrorsWithoutPattern { msgs, .. }]
                 if **code == "prefix::E0308" && code.line().get() == 3 && msgs.len() == 1 => {}
